api-server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
boto3==1.38.12
psycopg2-binary==2.9.10
pytesseract==0.3.13
Pillow==11.2.1
pdf2image==1.17.0
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.9.1
